+++ src/lib/components/SceneProviders.svelte
@@ -1,22 +1,17 @@
 <script lang="ts">
 	import type { Snippet } from 'svelte'
 
-	import { provideHierarchy } from '$lib/ecs'
+	import { provideHierarchy, provideWorldMatrix } from '$lib/ecs'
 	import { provide3DModels } from '$lib/hooks/use3DModels.svelte'
 	import { provideArmClient } from '$lib/hooks/useArmClient.svelte'
 	import { provideArmKinematics } from '$lib/hooks/useArmKinematics.svelte'
 	import { provideConfigFrames } from '$lib/hooks/useConfigFrames.svelte'
-	import {
-		type CameraPose,
-		provideCameraControls,
-		provideTransformControls,
-	} from '$lib/hooks/useControls.svelte'
-	import { provideDrawAPI } from '$lib/hooks/useDrawAPI.svelte'
-	import { provideDrawService } from '$lib/hooks/useDrawService.svelte'
+	import { provideTransformControls } from '$lib/hooks/useControls.svelte'
 	import { provideFrameEditSession } from '$lib/hooks/useFrameEditSession.svelte'
 	import { provideFramelessComponents } from '$lib/hooks/useFramelessComponents.svelte'
 	import { provideFrames } from '$lib/hooks/useFrames.svelte'
 	import { provideGeometries } from '$lib/hooks/useGeometries.svelte'
+	import { provideInheritedInvisible } from '$lib/hooks/useInheritedInvisible.svelte'
 	import { provideLinkedEntities } from '$lib/hooks/useLinked.svelte'
 	import { provideLogs } from '$lib/hooks/useLogs.svelte'
 	import { usePartID } from '$lib/hooks/usePartID.svelte'
@@ -30,23 +25,22 @@
 	import { provideOrigin } from './xr/useOrigin.svelte'
 
 	interface Props {
-		cameraPose?: CameraPose
 		children: Snippet<[{ focus: boolean }]>
 	}
 
-	let { cameraPose, children }: Props = $props()
+	let { children }: Props = $props()
 
 	const partID = usePartID()
 
-	provideCameraControls(() => cameraPose)
 	provideTransformControls()
 	provideLogs()
 
 	provideHierarchy()
+	provideWorldMatrix()
+	provideInheritedInvisible()
 	provideOrigin()
-	provideDrawAPI()
+
 	provideRelationships()
-	provideDrawService()
 
 	provideResourceByName(() => partID.current)
 	provideConfigFrames()