+++ src/lib/components/overlay/dashboard/Button.svelte
@@ -10,7 +10,7 @@
 		description: string
 		hotkey?: string
 		class?: ClassValue | null | undefined
-		tooltipLocation?: 'bottom' | 'right'
+		tooltipLocation?: 'bottom' | 'right' | 'left' | 'top'
 		onclick?: MouseEventHandler<HTMLButtonElement> | null | undefined
 	}
 
@@ -33,13 +33,13 @@
 	<label
 		class={[
 			className,
-			'relative block border',
-			active ? 'border-gray-5 text-gray-8 z-4 bg-white' : 'bg-light border-medium text-disabled',
+			'relative block rounded-md border',
+			active ? 'z-4 border-[#666] bg-[#666] text-white' : 'border-gray-5 text-gray-8 bg-white',
 		]}
 		aria-describedby={tooltipID}
 	>
 		<button
-			class="p-1.5"
+			class=" p-1.5"
 			role="radio"
 			aria-label={description}
 			aria-checked={active}