+++ src/lib/components/overlay/settings/Settings.svelte
@@ -62,7 +62,7 @@
 <Portal id="dashboard">
 	<fieldset>
 		<DashboardButton
-			active
+			active={isOpen.current}
 			icon="cog"
 			description="Settings"
 			onclick={() => {
@@ -278,10 +278,6 @@
 
 {#snippet Stats()}
 	<div class="flex w-full flex-col gap-2.5 text-xs">
-		<label class="flex items-center justify-between gap-2">
-			Query devtools <Switch bind:on={settings.current.enableQueryDevtools} />
-		</label>
-
 		<label class="flex items-center justify-between gap-2">
 			Render stats <Switch bind:on={settings.current.renderStats} />
 		</label>