+++ src/lib/plugins/index.ts
@@ -1,0 +1,18 @@
+// Selection
+export { default as SelectionTool } from './Selection/SelectionTool.svelte'
+export * as selectionTraits from './Selection/traits'
+export * as selectionRelations from './Selection/relations'
+export { useSelectionPlugin } from './Selection/useSelectionPlugin.svelte'
+
+export { default as MeasureTool } from './MeasureTool/MeasureTool.svelte'
+
+// DrawService
+export { default as DrawService } from './DrawService/DrawService.svelte'
+
+// Skybox
+export { default as Skybox } from './Skybox/Skybox.svelte'
+
+// Debug
+export { default as Debug } from './Debug/Debug.svelte'
+
+export { default as XR } from './XR/XR.svelte'