+++ src/routes/select/+page.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-	import { PCD, SelectionTool } from '$lib'
+	import { PCD } from '$lib'
+	import { SelectionTool } from '$lib/plugins'
 	import { createRandomPcdBinary } from '$lib/test/createRandomPcdBinary'
 
 	import SelectionOverlay from '../lib/components/selectDemoPage/SelectionOverlay.svelte'
@@ -11,6 +12,15 @@
 
 {#await createRandomPcdBinary(10_000, 1) then data}
 	<PCD
+		name="Points 1"
+		{data}
+		interactionLayers={['selectTool']}
+	/>
+{/await}
+
+{#await createRandomPcdBinary(1_000, 1) then data}
+	<PCD
+		name="Points 2"
 		{data}
 		interactionLayers={['selectTool']}
 	/>